Backend + Group + Scratchpad: Rename FloatStates to WindowStates
This makes it less confusing how we want to deal with the state
now. NOT_FLOATING is renamed to TILED as well.

The idea is that floating is that minimized, maximized, fullscreen,
etc... windows will not be considered 'floating' anymore. This makes
the code easier to reason and hack on